requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
anthropic==0.40.0
litellm>=1.40.0
reportlab==4.2.0
python-dotenv==1.0.1
python-multipart==0.0.9
jinja2==3.1.4
